pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maker-courseware"
version = "0.1.0"
description = "Local courseware generator driven by OpenAI Agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Codex" }
]
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "fastapi>=0.115,<1",
  "jinja2>=3.1,<4",
  "openai-agents>=0.2,<1",
  "pypdf>=5,<6",
  "python-docx>=1.1,<2",
  "python-multipart>=0.0.9,<1",
  "python-pptx>=0.6.23,<1",
  "uvicorn>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1",
  "pytest>=8,<9",
]

[project.scripts]
maker = "maker.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
maker = ["templates/*.html", "static/*.css"]
